requirements.txt
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-osc==1.8.0
python-rtmidi==1.4.9